Specs: Silence deprecation warning
Before:

> Randomized with seed 53114
> .........................................................................................................................................[WARNING] Using :ruby format when importing rates is potentially unsafe and might lead to remote code execution via Marshal.load deserializer. Consider using safe alternatives such as :json and :yaml.
> ...........................................................................................................................................................................................................................................................................................................................................................................................

After:

> Randomized with seed 30224
> ....................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................

Also moved a few RSpec configuration into their own support files in order to have a more readable RSpec configuration.
