+++ spec/spec_helper.rb
@@ -16,8 +16,6 @@ require "money"
 
 Dir['./spec/support/**/*.rb'].sort.each { |f| require f }
 
-I18n.enforce_available_locales = false
-
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -116,37 +114,4 @@ RSpec.configure do |config|
   config.before(:each) do
     Money.default_currency = Money::Currency.new("USD")
   end
-end
-
-def reset_i18n
-  I18n.backend = I18n::Backend::Simple.new
-end
-
-RSpec.shared_context "with infinite precision", :default_infinite_precision_true do
-  before do
-    Money.default_infinite_precision = true
-  end
-
-  after do
-    Money.default_infinite_precision = false
-  end
-end
-
-class Money
-  module Warning
-    def warn(message); end
-  end
-end
-
-class Money
-  include Warning
-  extend Warning
-end
-
-class Money::LocaleBackend::Base
-  include Money::Warning
-end
-
-class Money::FormattingRules
-  include Money::Warning
 end
